Finish refactor, prior to testing
