+++ src/matchbox/common/db.py
@@ -3,20 +3,8 @@ from typing import TYPE_CHECKING, Any, Literal, TypeVar, Union, overload
 import connectorx as cx
 import pyarrow as pa
 from pandas import DataFrame
-from pydantic import (
-    BaseModel,
-    ConfigDict,
-    Field,
-    SecretStr,
-    model_validator,
-)
-from sqlalchemy import (
-    create_engine,
-)
-from sqlalchemy import text as sqltext
 from sqlalchemy.engine import Engine
 from sqlalchemy.engine.url import URL
-from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.sql.selectable import Select
 
 if TYPE_CHECKING:
@@ -29,117 +17,7 @@ ReturnTypeStr = Literal["arrow", "pandas", "polars"]
 T = TypeVar("T")
 
 
-class Match(BaseModel):
-    """A match between primary keys in the Matchbox database."""
-
-    cluster: int | None
-    source: str
-    source_id: set[str] = Field(default_factory=set)
-    target: str
-    target_id: set[str] = Field(default_factory=set)
-
-    @model_validator(mode="after")
-    def found_or_none(self) -> "Match":
-        if self.target_id and not (self.source_id and self.cluster):
-            raise ValueError(
-                "A match must have sources and a cluster if target was found."
-            )
-        if self.cluster and not self.source_id:
-            raise ValueError("A match must have source if cluster is set.")
-        return self
-
-
-class Probability(BaseModel):
-    """A probability of a match in the Matchbox database.
-
-    A probability describes the likelihood of a match between two clusters.
-    """
-
-    id: bytes
-    left: int
-    right: int
-    probability: float = Field(default=None, ge=0, le=1)
-
-
-class Cluster(BaseModel):
-    """A cluster of data in the Matchbox database.
-
-    A cluster describes a single entity resolved at the specified probability
-    threshold or higher.
-    """
-
-    parent: bytes
-    children: set[bytes]
-    threshold: float = Field(default=None, ge=0, le=1)
-
-
-class SourceWarehouse(BaseModel):
-    """A warehouse where source data for datasets in Matchbox can be found."""
-
-    model_config = ConfigDict(
-        populate_by_name=True,
-        extra="forbid",
-        arbitrary_types_allowed=True,
-    )
-
-    alias: str
-    db_type: str
-    user: str
-    password: SecretStr
-    host: str
-    port: int
-    database: str
-    _engine: Engine | None = None
-
-    @property
-    def engine(self) -> Engine:
-        if self._engine is None:
-            connection_string = f"{self.db_type}://{self.user}:{self.password.get_secret_value()}@{self.host}:{self.port}/{self.database}"
-            self._engine = create_engine(connection_string)
-            self.test_connection()
-        return self._engine
-
-    def __eq__(self, other):
-        if not isinstance(other, SourceWarehouse):
-            return False
-        return (
-            self.alias == other.alias
-            and self.db_type == other.db_type
-            and self.user == other.user
-            and self.password == other.password
-            and self.host == other.host
-            and self.port == other.port
-            and self.database == other.database
-        )
-
-    def test_connection(self):
-        try:
-            with self.engine.connect() as connection:
-                connection.execute(sqltext("SELECT 1"))
-        except SQLAlchemyError:
-            self._engine = None
-            raise
-
-    @classmethod
-    def from_engine(cls, engine: Engine, alias: str | None = None) -> "SourceWarehouse":
-        """Create a SourceWarehouse instance from an SQLAlchemy Engine object."""
-        url = engine.url
-
-        warehouse = cls(
-            alias=alias or url.database,
-            db_type=url.drivername,
-            user=url.username,
-            password=url.password,
-            host=url.host,
-            port=url.port or 0,
-            database=url.database,
-        )
-        _ = warehouse.engine
-
-        return warehouse
-
-
-def convert_large_binary_to_binary(table: pa.Table) -> pa.Table:
+def _convert_large_binary_to_binary(table: pa.Table) -> pa.Table:
     """Converts Arrow large_binary fields to binary type."""
     new_fields = []
     for field in table.schema:
@@ -202,6 +80,6 @@ def sql_to_df(
     result = cx.read_sql(conn=url, query=sql_query, return_type=return_type)
 
     if return_type == "arrow":
-        return convert_large_binary_to_binary(table=result)
+        return _convert_large_binary_to_binary(table=result)
 
     return result
